Extracted ElasticSearchClient interface and renamed implementation
